+++ activesupport/lib/active_support/notifications.rb
@@ -2,6 +2,7 @@
 
 require "active_support/notifications/instrumenter"
 require "active_support/notifications/fanout"
+require "active_support/notifications/event_assertions"
 
 module ActiveSupport
   # = \Notifications