+++ activesupport/CHANGELOG.md
@@ -17,6 +17,10 @@
 
     *Jean Boussier*
 
+*   Add `ActiveSupport::Notifications::EventAssertions` module to help with testing events.
+
+    *Nicholas La Roux*, *Yishu See*
+
 ## Rails 8.0.0.beta1 (September 26, 2024) ##
 
 *   Include options when instrumenting `ActiveSupport::Cache::Store#delete` and `ActiveSupport::Cache::Store#delete_multi`.